+++ src/tests/mod.rs
@@ -115,7 +115,10 @@ impl Nitrocli {
       new_admin_pin: self.new_admin_pin.clone(),
       new_user_pin: self.new_user_pin.clone(),
       password: self.password.clone(),
-      no_cache: true,
+      config: crate::config::Config {
+        no_cache: true,
+        ..Default::default()
+      },
     };
 
     (f(ctx, args), stdout, stderr)